Terminate le ultime modifiche comportamentali del progetto: Disconnessione istantanea quando si perde la connessione col server e quando si viene bannati; Eliminato il TextEditor dal client che non veniva più utilizzato; Inserito Controllo sulla lunghezza minima dell'Username (Minimo 3 caratteri); Modificato il trigger dell'inserimento sfida per fare in modo che non conti le partite dove un giocatore si è arreso; Risolto un bug che faceva comparire il messaggio d'errore sbagliato nella fase di login; Risolto Spam continuo del messaggio STORICO_OK da parte del Server; Spostati i testi e i report nel progetto Server (L'accesso è consentito solo a lui); Modificato il nome dei file properties per matchare il formato richiesto dal professore; Aggiunta schermata d'attesa avversario per i giocatori che hanno inviato la risposta; Spostato il messaggio d'errore legato al server nella pagina di login in modo da essere maggiormente visibile.

+++ GuessTheWord(Client)/src/main/java/model/utility/Sfida.java
@@ -1,29 +1,21 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
 package model.utility;
 
 import java.io.Serializable;
 
-/**
- *
- * @author euppa
- */
 public class Sfida implements Serializable{
     
     private String titoloTesto;
     private int idDocumento;
-    private double durata;
-    private double tRisposta1;
-    private double tRisposta2;
+    private int durata;
+    private int tRisposta1;
+    private int tRisposta2;
     private int id1;
     private int id2;
     private String oppUsername;
     private String risultato;
     private String soluzione;
 
-    public Sfida(int idDocumento, double durata, double tRisposta1, double tRisposta2, int id1, int id2, String oppUsername, String risultato, String soluzione) {
+    public Sfida(int idDocumento, int durata, int tRisposta1, int tRisposta2, int id1, int id2, String oppUsername, String risultato, String soluzione) {
         this.idDocumento = idDocumento;
         this.durata = durata;
         this.tRisposta1 = tRisposta1;
@@ -43,15 +35,15 @@ public class Sfida implements Serializable{
         return idDocumento;
     }
 
-    public double getDurata() {
+    public int getDurata() {
         return durata;
     }
 
-    public double getTRisposta1() {
+    public int getTRisposta1() {
         return tRisposta1;
     }
 
-    public double getTRisposta2() {
+    public int getTRisposta2() {
         return tRisposta2;
     }
 
@@ -83,15 +75,15 @@ public class Sfida implements Serializable{
         this.idDocumento = idDocumento;
     }
 
-    public void setDurata(double durata) {
+    public void setDurata(int durata) {
         this.durata = durata;
     }
 
-    public void settRisposta1(double tRisposta1) {
+    public void settRisposta1(int tRisposta1) {
         this.tRisposta1 = tRisposta1;
     }
 
-    public void settRisposta2(double tRisposta2) {
+    public void settRisposta2(int tRisposta2) {
         this.tRisposta2 = tRisposta2;
     }
 
@@ -120,13 +112,4 @@ public class Sfida implements Serializable{
         return titoloTesto + ", " + idDocumento + ", " + durata + ", " + tRisposta1 + ", " + 
                 tRisposta2 + ", " + id1 + ", " + id2 + ", " + oppUsername + ", " + risultato + ", " + soluzione;
     }
-    
-    
-    
-    
-    
-    
-    
-    
-    
 }
